backend/pyproject.toml
[project]
name = "solaire"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "faker>=37.4.0",
    "fastapi[standard]>=0.115.14",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.26",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.5.0",
    "ollama>=0.5.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "tomli>=2.2.1",
    "weaviate-client>=4.0.0",
]
